Move VCL API to utils
