+++ src/plugins/intel_npu/src/utils/include/intel_npu/utils/vcl/vcl.h
@@ -249,21 +249,10 @@ VCL_APIEXPORT vcl_result_t VCL_APICALL vclExecutableCreate(vcl_compiler_handle_t
                                                            vcl_executable_desc_t desc,
                                                            vcl_executable_handle_t* executable);
 
-DEPRECATED typedef struct __vcl_allocator_t {
-    uint8_t* (*allocate)(uint64_t);
-    void (*deallocate)(uint8_t*);
-} vcl_allocator_t;
-
 typedef struct __vcl_allocator2_t {
     uint8_t* (*allocate)(struct __vcl_allocator2_t*, uint64_t);
     void (*deallocate)(struct __vcl_allocator2_t*, uint8_t*);
 } vcl_allocator2_t;
-
-DEPRECATED VCL_APIEXPORT vcl_result_t VCL_APICALL vclAllocatedExecutableCreate(vcl_compiler_handle_t compiler,
-                                                                               vcl_executable_desc_t desc,
-                                                                               const vcl_allocator_t* allocator,
-                                                                               uint8_t** blobBuffer,
-                                                                               uint64_t* blobSize);
 
 VCL_APIEXPORT vcl_result_t VCL_APICALL vclAllocatedExecutableCreate2(vcl_compiler_handle_t compiler,
                                                                      vcl_executable_desc_t desc,
